+++ src/ui/localstorage.ts
@@ -78,7 +78,12 @@ export const setPreferenceMachineName = (name: MACHINE_NAME = "APPLE2EE") => {
   passSetMachineName(name)
 
   const slotConfig = getPreferenceSlotConfig()
-  const newSlot3 = name === "APPLE2P" ? "none" : "aux"
+  let newSlot3 = slotConfig[3]
+  if (name === "APPLE2P") {
+    if (newSlot3 !== "none" && newSlot3 !== "videoterm") newSlot3 = "videoterm"
+  } else {
+    if (newSlot3 !== "none" && newSlot3 !== "aux") newSlot3 = "aux"
+  }
   if (slotConfig[3] !== newSlot3) {
     slotConfig[3] = newSlot3
     setPreferenceSlotConfig(slotConfig)